+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This documents the changes from release to release.
 
 ### 0.0.10
 
+
+For issue https://github.com/MCLConsortium/mcl.sickbay/issues/1:
+
 -   On `ClinicalCore`:
     -   The `race` attribute is now a 1-to-many mapping to `CoreRace` via `core_races`
     -   The `type_tobacco_used` is now a 1-to-many mapping to `CoreTobacco` via `core_tobaccos`
@@ -25,6 +28,10 @@ This documents the changes from release to release.
     -   Now it's completely filled out with the `v0` prostate common data elements with its numerous controlled vocabularies
 -   Expanded enumerations: `ClinicalMStage7`, `TStage7`, `ClinicalNStage7`, `GroupStage7`, `MarginalStatus`
 -   New enumerations, far too many to enumerate 😏
+
+For issue https://github.com/MCLConsortium/mcl.sickbay/issues/4:
+
+-   All fields in `LabCASMetadata` are now `String`.
 
 
 ### 0.0.9